Fixed db
The user register now posts the data to mongo
