+++ App/backend/routes/user.js
@@ -1,11 +1,10 @@
 const express = require('express');
-const { registerUser } = require('../controllers/userContoller');
+const { registerUser, loginUser } = require('../controllers/userController');
 
 const router = express.Router();
 
+// Routes 
 router.post('/register', registerUser);
-
-// need to create model 
 router.post('/login', loginUser);
 
 module.exports = router;
