+++ App/package.json
@@ -13,6 +13,7 @@
     "@expo/metro-runtime": "~4.0.1",
     "@react-navigation/native-stack": "^7.2.1",
     "@react-navigation/stack": "^7.1.2",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "expo": "~52.0.37",
     "expo-status-bar": "~2.0.1",
